// apps/desktop/src/types/artemis.ts
export interface RpcEvent<T = unknown> {
  payload?: T
  profile?: string
  session_id?: string
  type: string
}